# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LineageBridge API
  description: OpenLineage-compatible REST API that exposes Confluent Cloud stream
    lineage for integration with data catalogs (Databricks UC, AWS Glue, Google Data
    Lineage, and more).
  version: 0.3.0
paths:
...
  /api/v1/graphs:
    get:
      tags:
      - graphs
      summary: List Graphs
      description: List all in-memory graphs.
      operationId: list_graphs_api_v1_graphs_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GraphSummary'
                type: array
                title: Response List Graphs Api V1 Graphs Get
      security:
      - APIKeyHeader: []
    post:
      tags:
      - graphs
      summary: Create Graph
      description: Create a new empty graph.
      operationId: create_graph_api_v1_graphs_post
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Create Graph Api V1 Graphs Post
      security:
      - APIKeyHeader: []
  /api/v1/graphs/{graph_id}:
    get:
      tags:
      - graphs
      summary: Get Graph
      description: Get a full graph with nodes, edges, and stats.
      operationId: get_graph_api_v1_graphs__graph_id__get
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Graph Api V1 Graphs  Graph Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - graphs
      summary: Delete Graph
      operationId: delete_graph_api_v1_graphs__graph_id__delete
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Delete Graph Api V1 Graphs  Graph Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graphs/{graph_id}/import:
    post:
      tags:
      - graphs
      summary: Import Graph
      description: Import nodes and edges from JSON into an existing graph.
      operationId: import_graph_api_v1_graphs__graph_id__import_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Import Graph Api V1 Graphs  Graph Id  Import Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graphs/{graph_id}/export:
    get:
      tags:
      - graphs
      summary: Export Graph
      description: Export graph as native LineageGraph JSON.
      operationId: export_graph_api_v1_graphs__graph_id__export_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Export Graph Api V1 Graphs  Graph Id  Export Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graphs/confluent/view:
    get:
      tags:
      - graphs
      summary: Confluent View
      description: 'Confluent-only lineage view across all graphs.


        Returns OpenLineage events for Confluent-system nodes only,

        excluding catalog nodes (UC, Glue, Google).'
      operationId: confluent_view_api_v1_graphs_confluent_view_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Confluent View Api V1 Graphs Confluent View Get
      security:
      - APIKeyHeader: []
  /api/v1/graphs/enriched/view:
    get:
      tags:
      - graphs
      summary: Enriched View
      description: 'Full enriched lineage view across all graphs.


        Includes Confluent nodes plus all catalog nodes (UC, Glue, Google).

        Optionally filter by system types.'
      operationId: enriched_view_api_v1_graphs_enriched_view_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: systems
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Comma-separated system filter (e.g. confluent,databricks)
          title: Systems
        description: Comma-separated system filter (e.g. confluent,databricks)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Enriched View Api V1 Graphs Enriched View Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graphs/{graph_id}/nodes:
    get:
      tags:
      - graphs
      summary: List Nodes
      description: List nodes with optional filters.
      operationId: list_nodes_api_v1_graphs__graph_id__nodes_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      - name: type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/NodeType'
          - type: 'null'
          title: Type
      - name: system
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/SystemType'
          - type: 'null'
          title: System
      - name: env_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Env Id
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Nodes Api V1 Graphs  Graph Id  Nodes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - graphs
      summary: Add Node
      operationId: add_node_api_v1_graphs__graph_id__nodes_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineageNode'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Add Node Api V1 Graphs  Graph Id  Nodes Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graphs/{graph_id}/nodes/{node_id}:
    get:
      tags:
      - graphs
      summary: Get Node
      operationId: get_node_api_v1_graphs__graph_id__nodes__node_id__get
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageNode'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graphs/{graph_id}/edges:
    post:
      tags:
      - graphs
      summary: Add Edge
      operationId: add_edge_api_v1_graphs__graph_id__edges_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineageEdge'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Add Edge Api V1 Graphs  Graph Id  Edges Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - graphs
      summary: List Edges
      operationId: list_edges_api_v1_graphs__graph_id__edges_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Edges Api V1 Graphs  Graph Id  Edges Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graphs/{graph_id}/query/upstream/{node_id}:
    get:
      tags:
      - graphs
      summary: Query Upstream
      operationId: query_upstream_api_v1_graphs__graph_id__query_upstream__node_id__get
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      - name: hops
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 3
          title: Hops
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Query Upstream Api V1 Graphs  Graph Id  Query Upstream  Node
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graphs/{graph_id}/query/downstream/{node_id}:
    get:
      tags:
      - graphs
      summary: Query Downstream
      operationId: query_downstream_api_v1_graphs__graph_id__query_downstream__node_id__get
      security:
      - APIKeyHeader: []
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          title: Graph Id
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      - name: hops
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 3
          title: Hops
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Query Downstream Api V1 Graphs  Graph Id  Query Downstream  Node
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EdgeType:
      type: string
      enum:
      - produces
      - consumes
      - transforms
      - materializes
      - has_schema
      - member_of
      title: EdgeType
      description: Types of edges (relationships) in the lineage graph.
    GraphSummary:
      properties:
        graph_id:
          type: string
          title: Graph Id
        node_count:
          type: integer
          title: Node Count
        edge_count:
          type: integer
          title: Edge Count
        pipeline_count:
          type: integer
          title: Pipeline Count
        created_at:
          type: string
          format: date-time
          title: Created At
        last_modified:
          type: string
          format: date-time
          title: Last Modified
      type: object
      required:
      - graph_id
      - node_count
      - edge_count
      - pipeline_count
      - created_at
      - last_modified
      title: GraphSummary
      description: Lightweight graph info for listing.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LineageEdge:
      properties:
        src_id:
          type: string
          title: Src Id
        dst_id:
          type: string
          title: Dst Id
        edge_type:
          $ref: '#/components/schemas/EdgeType'
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          description: 1.0 = deterministic, <1.0 = inferred
          default: 1.0
        attributes:
          additionalProperties: true
          type: object
          title: Attributes
        first_seen:
          type: string
          format: date-time
          title: First Seen
        last_seen:
          type: string
          format: date-time
          title: Last Seen
      type: object
      required:
      - src_id
      - dst_id
      - edge_type
      title: LineageEdge
      description: A directed edge in the lineage graph representing a data flow relationship.
    LineageNode:
      properties:
        node_id:
          type: string
          title: Node Id
          description: 'Unique identifier. Format: {system}:{type}:{env_id}:{qualified_name}'
        system:
          $ref: '#/components/schemas/SystemType'
        node_type:
          $ref: '#/components/schemas/NodeType'
        qualified_name:
          type: string
          title: Qualified Name
        display_name:
          type: string
          title: Display Name
        environment_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Environment Id
        environment_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Environment Name
        cluster_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cluster Id
        cluster_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Cluster Name
        attributes:
          additionalProperties: true
          type: object
          title: Attributes
        tags:
          items:
            type: string
          type: array
          title: Tags
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        first_seen:
          type: string
          format: date-time
          title: First Seen
        last_seen:
          type: string
          format: date-time
          title: Last Seen
      type: object
      required:
      - node_id
      - system
      - node_type
      - qualified_name
      - display_name
      title: LineageNode
      description: A node in the lineage graph representing a data asset or processing
        step.
    NodeType:
      type: string
      enum:
      - kafka_topic
      - connector
      - ksqldb_query
      - flink_job
      - tableflow_table
      - uc_table
      - glue_table
      - google_table
      - schema
      - external_dataset
      - consumer_group
      title: NodeType
      description: Types of nodes in the lineage graph.
...
    SystemType:
      type: string
      enum:
      - confluent
      - databricks
      - aws
      - google
      - external
      title: SystemType
      description: Source systems that contribute nodes to the graph.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key